+++ src/lib/helpers.test.ts
@@ -4,6 +4,8 @@
 import { CapabilityExport } from "./types";
 import { createRBACMap, addVerbIfNotExists } from "./helpers";
 import { expect, describe, test, jest, beforeEach, afterEach } from "@jest/globals";
+import { parseTimeout } from "./helpers";
+import * as commander from "commander";
 import { promises as fs } from "fs";
 import {
   createDirectoryIfNotExists,
@@ -823,5 +825,29 @@ describe("namespaceDeploymentsReady", () => {
     expect(result).toBe(expected);
 
     expect(mockK8s).toHaveBeenCalledTimes(1);
+  });
+});
+
+describe("parseTimeout", () => {
+  const PREV = "a";
+  test("should return a number when a valid string number between 1 and 30 is provided", () => {
+    expect(parseTimeout("5", PREV)).toBe(5);
+    expect(parseTimeout("1", PREV)).toBe(1);
+    expect(parseTimeout("30", PREV)).toBe(30);
+  });
+
+  test("should throw an InvalidArgumentError for non-numeric strings", () => {
+    expect(() => parseTimeout("abc", PREV)).toThrow(commander.InvalidArgumentError);
+    expect(() => parseTimeout("", PREV)).toThrow(commander.InvalidArgumentError);
+  });
+
+  test("should throw an InvalidArgumentError for numbers outside the 1-30 range", () => {
+    expect(() => parseTimeout("0", PREV)).toThrow(commander.InvalidArgumentError);
+    expect(() => parseTimeout("31", PREV)).toThrow(commander.InvalidArgumentError);
+  });
+
+  test("should throw an InvalidArgumentError for numeric strings that represent floating point numbers", () => {
+    expect(() => parseTimeout("5.5", PREV)).toThrow(commander.InvalidArgumentError);
+    expect(() => parseTimeout("20.1", PREV)).toThrow(commander.InvalidArgumentError);
   });
 });
